% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_freqsum_table.r
\name{fun_freqsum_table}
\alias{fun_freqsum_table}
\title{Frequency summary report, a MJP team defined function}
\usage{
fun_freqsum_table(dataset, path_report_output, x = 20, sort_by = "COUNT",
  sort_descending = T)
}
\arguments{
\item{dataset}{a dataset to run the summary}

\item{path_report_output}{a path to store the summary output, have to be specified as a .csv file.}

\item{x}{number of categories shown in the summary table, default = 20.}

\item{sort_by}{sort summary table by "VARIABLE" = category lable or "COUNT" = freqency}

\item{sort_descending}{show the result in acsending or descending order, default = TRUE, in descending order.}
}
\value{
A .csv file has the frequency summary table for each of the variable.
}
\description{
Frequency summary report, a MJP team defined function
}
\examples{
    fun_freqsum_table(my_df , "C:\\\\Users\\\\tlin\\\\Dropbox\\\\Richie\\\\RFiles\\\\summary.csv")
    fun_freqsum_table(my_df , "C:\\\\Users\\\\tlin\\\\Dropbox\\\\Richie\\\\RFiles\\\\summary.csv", x = 50,  sort_by = "VARIABLE", sort_descending = T)
}
