% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_livefeed_dd_overview.r
\name{fun_livefeed_dd_overview}
\alias{fun_livefeed_dd_overview}
\title{Auto-generate the Overview table of MOCJ livefeed data dictionary}
\usage{
fun_livefeed_dd_overview(dataset_nm_keyword, db_connect = F,
  db_username = as.character(), db_password = as.character())
}
\arguments{
\item{dataset_nm_keyword}{String, The keyword which will be used to search for the dataset}

\item{db_connect}{T/F value, indicate whether the Database info (data type and length in DB) is desired.}

\item{db_username}{The username for logging into MOCJ livefeed database}

\item{db_password}{The password for logging into MOCJ livefeed database}
}
\value{
A data frame containning the following information of each variable in all tables: Table name, Column index, Variable name, Variable description (blank), PII indecater(Blank), Data type in database, length in database, data type in R, number of rows, missing count and missing percentage.
}
\description{
Auto-generate the Overview table of MOCJ livefeed data dictionary
}
\details{
This function will search for the dataset with the given keyword in the global enviroment. Thus, loading all of the dataset into the memory and naming them with the same exact keyword is requried. \n The variable description and PII indecator in the output are used as place holder, user has to either link with other references or manually input.
}
\examples{
    fun_freqsum_table(my_df , "C:\\\\Users\\\\tlin\\\\Dropbox\\\\Richie\\\\RFiles\\\\summary.csv")
    fun_freqsum_table(my_df , "C:\\\\Users\\\\tlin\\\\Dropbox\\\\Richie\\\\RFiles\\\\summary.csv", x = 50,  sort_by = "VARIABLE", sort_descending = T)
}
