% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_logic_assg.r
\name{fun_logic_assg}
\alias{fun_logic_assg}
\title{Logical assignment, a MJP team defined function}
\usage{
fun_logic_assg(x, reference, typo_col = "TYPO", replace_col = "REPLACEMENT")
}
\arguments{
\item{x}{A input variable/ column/ vector of strings to perform the logical assignment on}

\item{reference}{A reference table (lookup table) having at least 2 column for matching the typo and returning the replacement}

\item{typo_col}{The name ( as a "character") of the column in the reference table that allows the function to match with the input column "x"}

\item{replace_col}{The name (as a "character) of the column in the reference table that contains the coresponding values to be returned.}
}
\value{
A vector (character) that contains the replacement values from the reference table
}
\description{
Logical assignment, a MJP team defined function
}
