% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameCol.r
\name{nameCol}
\alias{nameCol}
\title{Vertically displayed column name function, a MJP team defined function}
\usage{
nameCol(data = mtcars)
}
\arguments{
\item{data}{Input dataset to extract column.}
}
\value{
A matrix displaying the columnor variable names vertically
}
\description{
Vertically displayed column name function, a MJP team defined function
}
\examples{
    nameCol(mtcars)
}
