% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqsum.r
\name{freqsum}
\alias{freqsum}
\title{Frequency summary, a MJP team defined function}
\usage{
freqsum(InputCol, x = 20, sort_by = "COUNT", sort_descending = T,
  big_mark = NA, round_percent = 4)
}
\arguments{
\item{InputCol}{Input column or vector, will be convert into factor.}

\item{x}{number of categories shown in the summary table, default = 20.}

\item{sort_by}{sort summary table by "VARIABLE" = category lable or "COUNT" = freqency}

\item{sort_descending}{show the result in acsending or descending order, default = TRUE, in descending order.}

\item{big_mark}{adding seperator in the big numbers. if you dont like the scientific notation when using this on large numbers, please change your globle enviromental variable with 'option(scipen = 999)' .}

\item{round_percent}{the accuracy of the percentage in the summary table of fator variable, 4 digits by default.}
}
\value{
A frequency summary table with percentage.
}
\description{
Frequency summary, a MJP team defined function
}
\examples{
    freqsum(my_df$col1)
    freqsum(my_df$col2, x = 50, sort_by = "VARIABLE", sort_descending = T)
}
